outsourced theming control in service
