+++ frontend/src/app/core/components/sidenav/sidenav.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Observable } from 'rxjs';
 import { map, shareReplay } from 'rxjs/operators';
+import { ThemesService } from '../../services/themes/themes.service';
 
 @Component({
   selector: 'app-sidenav',
@@ -16,24 +17,5 @@ export class SidenavComponent {
       shareReplay()
     );
 
-  constructor(private breakpointObserver: BreakpointObserver) {}
-
-  ngOnInit() {
-    const userPrefersLocal = localStorage.getItem('prefers-color-scheme'); //=== 'Light' ? false : true;
-    const userPrefersLightBrowser = window.matchMedia && window.matchMedia('(prefers-color-scheme: light)').matches;
-    if (userPrefersLocal) {
-      this.isDarkTheme = userPrefersLocal === 'light' ? false : true;
-    } else {
-      this.isDarkTheme = !userPrefersLightBrowser
-    }
-  }
-
-  toggleTheme() {
-    this.isDarkTheme = !this.isDarkTheme;
-    this.storeThemeSelection();
-  }
-
-  storeThemeSelection() {
-    localStorage.setItem('prefers-color-scheme', this.isDarkTheme ? "dark" : "light")
-  }
+  constructor(private breakpointObserver: BreakpointObserver, public themesService: ThemesService) {}
 }
